2022/requirements.txt
requests>=2.28.1
bs4>=0.0.1
lxml>=4.9.1
beautifulsoup4~=4.11.1
mysql-connector-python~=8.0.29
urllib3>=1.26.10
pandas>=1.4.3
openpyxl>=3.0.10